src/requirements.txt
discord.py
yt-dlp
ffmpeg
asyncio
PyNaCl
youtube-search-python